Refactor environment variable mock fixtures in tests for improved clarity and reuse
